+++ dpppt-backend-sdk/dpppt-backend-sdk-data/src/main/java/org/dpppt/backend/sdk/data/gaen/GAENDataService.java
@@ -62,7 +62,8 @@ public interface GAENDataService {
    *
    * @param keysSince
    * @param now
+   * @param visitedCountries
    * @return
    */
-  List<GaenKey> getSortedExposedSince(UTCInstant keysSince, UTCInstant now);
+  List<GaenKey> getSortedExposedSince(UTCInstant keysSince, UTCInstant now, List<String> visitedCountries);
 }