+++ dpppt-backend-sdk/dpppt-backend-sdk-ws/src/test/java/org/dpppt/backend/sdk/ws/insertmanager/MockDataSource.java
@@ -31,7 +31,8 @@ public class MockDataSource implements GAENDataService {
 
   @Override
   public List<GaenKey> getSortedExposedSince(
-      UTCInstant keysSince, UTCInstant now) { // TODO Auto-generated method stub
+      UTCInstant keysSince, UTCInstant now, List<String> visitedCountries) { 
+	// TODO Auto-generated method stub
     return null;
   }
 }